refactor(design-tokens): move feed classes into Admin\Feed namespace
